+++ polyglot/piranha/src/models/mod.rs
@@ -23,3 +23,4 @@ pub(crate) mod rule_graph;
 pub(crate) mod rule_store;
 pub(crate) mod scopes;
 pub(crate) mod source_code_unit;
+pub(crate) mod default_configs;
