+++ polyglot/piranha/src/models/unit_tests/scopes_test.rs
@@ -107,8 +107,7 @@ fn test_get_scope_query_positive() {
     rule_store.piranha_args(),
   );
 
-  let scope_query_method = ScopeGenerator::get_scope_query(
-    source_code_unit.clone(),
+  let scope_query_method = source_code_unit.get_scope_query(
     "Method",
     133,
     134,
@@ -136,7 +135,7 @@ fn test_get_scope_query_positive() {
   ));
 
   let scope_query_class =
-    ScopeGenerator::get_scope_query(source_code_unit, "Class", 133, 134, &mut rule_store);
+  source_code_unit.get_scope_query("Class", 133, 134, &mut rule_store);
   assert!(eq_without_whitespace(
     scope_query_class.as_str(),
     "(
@@ -170,5 +169,5 @@ fn test_get_scope_query_negative() {
     rule_store.piranha_args(),
   );
 
-  let _ = ScopeGenerator::get_scope_query(source_code_unit, "Method", 9, 10, &mut rule_store);
+  let _ = source_code_unit.get_scope_query("Method", 9, 10, &mut rule_store);
 }
